msm: mdss: mdp: secure buffer is unmapped without iommu attached
For secure buffers, it is a requirement that IOMMU is attached
before unmapping can be called, otherwise the scm call to unmap the
buffer will not be successful. In the current implementation it is
possible to unmap buffer while going into suspend without IOMMU being
attached which leads to secure playback failures in subsequent sessions.
Fix this by ensuring IOMMU is attached before buffer is freed/unmapped.

Change-Id: I29cdce834356b45f05fc75ec07d1c6fa0fcf2d69
Signed-off-by: Siva Yarravarapu <sivay@codeaurora.org>
Signed-off-by: Naseer Ahmed <naseer@codeaurora.org>